+++ src/MvcCore/Ext/Controllers/DataGrids/AgGrids/Options/Manager.ts
@@ -201,7 +201,7 @@ namespace MvcCore.Ext.Controllers.DataGrids.AgGrids.Options {
 			// how many pages to store in cache. default is undefined, which allows an infinite sized cache,
 			// pages are never purged. this should be set for large data to stop your browser from getting
 			// full of data
-			if (serverConfig.clientMaxRowsInCache > 0) {
+			if (serverConfig.clientCache && serverConfig.clientMaxRowsInCache > 0) {
 				this.agOptions.maxBlocksInCache = Math.round(serverConfig.clientMaxRowsInCache / serverConfig.clientRequestBlockSize);
 			} else {
 				this.agOptions.maxBlocksInCache = 1;