+++ changelog.md
@@ -1,2 +1,3 @@
-### Fixed
-- Commented back `X-Requested-With` header in column change request.
+### New
+- Automatic row id computation by server configuration.
+- Method `agGrid.UpdateRows(rowData: any[]): this` to update rows in both datasource types.